# Improve frontpercent performance