### **Overview**

Refactored GetTruncatedSwitchList to:
- Only fetch switches in the requested range
- Fetch switch members in bulk rather than one switch at a time

Important: The queries in the new GetSwitchMembersList method work best with the new indexes added to the schema. idx_switch_members_switch renders an index that was manually added to production obsolete (it covers the same column, but adds an include column). That existing index can  (and should) just be removed.

### **Performance results**

Tests were run on a database with the following characteristics:
- 740k total systems
- 300k total members
- 1.8mm total switches
- 4.3mm total switch_members records

I used a system with the following characteristics:
- 150 members
- 18k switches
- 900k switch_members records

**Before:**
pk;s fp - 16s
pk;s fp 1d - 0.137s
pk;s fp 1h - 0.145s

**After:**
pk;s fp - 0.915s (94% improvement)
pk;s fp 1d - 0.0426s (72% improvement)
pk;s fp 1h - 0.0074s (95% improvement)

### **Questions**

- How should we manage standard indexes going forward?
- What naming conventions should be used for indexes?